requirements.txt
annotated-types==0.7.0
anyio==4.7.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
decorator==4.4.2
distro==1.9.0
frozendict==2.4.6
gitignore_parser==0.1.11
greenlet==3.1.1
gTTS==2.2.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
jiter==0.8.0
lxml==5.3.0
moviepy==1.0.3
multitasking==0.0.11
numpy==2.1.3
openai==1.57.0
pandas==2.2.3
peewee==3.17.8
Pillow==9.5.0
platformdirs==4.3.6
playwright==1.49.0
proglog==0.1.10
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1
pyee==12.0.0
pygame==2.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
srt==3.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vosk==0.3.44
webencodings==0.5.1
websockets==14.1
yfinance==0.2.50
